Move binding logic to an initialise method
