+++ core/src/main/java/com/novoda/noplayer/player/PlayerFactory.java
@@ -5,7 +5,8 @@ import android.content.Context;
 import com.novoda.noplayer.Player;
 import com.novoda.noplayer.drm.DrmHandler;
 import com.novoda.noplayer.exoplayer.ExoPlayerTwoImpl;
-import com.novoda.noplayer.mediaplayer.AndroidMediaPlayerImplCreator;
+import com.novoda.noplayer.mediaplayer.AndroidMediaPlayerImpl;
+import com.novoda.noplayer.mediaplayer.AndroidMediaPlayerImplFactory;
 
 public class PlayerFactory {
 
@@ -41,8 +42,10 @@ public class PlayerFactory {
         }
     }
 
-    private Player createMediaPlayer() {
-        return new AndroidMediaPlayerImplCreator().create(context);
+    private AndroidMediaPlayerImpl createMediaPlayer() {
+        AndroidMediaPlayerImpl player = new AndroidMediaPlayerImplFactory().create(context);
+        player.initialise();
+        return player;
     }
 
     private static class UnableToCreatePlayerException extends RuntimeException {
